Added login auth middleware for Admin page and API

+++ pages/api/deleteProject.js
@@ -1,6 +1,9 @@
-import db from '/database'
+import db from '/functions/database'
+import isAdmin from '/functions/isAdmin'
 
 export default function handler(req, res) {
+  if (!isAdmin(req)) return res.status(403).send({ error: 'No permission' })
+
   db.serialize(() => {
     db.run(`DELETE FROM Projects WHERE id = (?1)`, { 1: req.body.id })
     res.status(200).json({})
